chore: move period selector into own directory
