+++ src/components/PeriodSelector/FixedPeriodFilter.js
