+++ src/components/PeriodSelector/InputLabel.js
