+++ src/components/PeriodSelector/modules/FixedPeriodsGenerator.js
