+++ src/components/PeriodSelector/modules/RelativePeriodsGenerator.js
