+++ src/components/PeriodSelector/modules/__tests__/FixedPeriodsGenerator.spec.js
