+++ src/components/PeriodSelector/modules/periodTypes.js
