+++ src/components/PeriodSelector/styles/InputLabel.style.js
