+++ src/components/PeriodSelector/styles/PeriodSelector.style.js
