+++ src/components/PeriodSelector/styles/PeriodTypeButton.style.js
